% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelEstimate.R
\name{kernelEstimate}
\alias{kernelEstimate}
\title{Kernel Estimate}
\usage{
kernelEstimate(x1, x2, h, lambda, kernelType)
}
\arguments{
\item{x1}{First data point}

\item{x2}{Second data point}

\item{h}{Bandwidth}

\item{lambda}{Lambda, only used when KernelType is flat}

\item{kernelType}{Type of the kernel to be used, either "gaussian" or "flat"}
}
\value{
Kernel value.
}
\description{
Kernel Estimate
}
\examples{

}
