% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdrags.R
\docType{package}
\name{Rdrags}
\alias{Rdrags}
\alias{Rdrags-package}
\alias{Rdrags}
\title{Rdrags: High-dimensional feature selection via Dense Relevant Attribute Group Selector}
\usage{
Rdrags(data, maxFeatures = 10, h = 0, k = 5, useKNN = TRUE,
  kernelType = "gaussian", eps = 1e-05, maxIter = 100, lambda = 0)
}
\arguments{
\item{data}{Data to process}

\item{maxFeatures}{Number of features to be selected.}

\item{h}{Bandwidth, will not be used when useKNN is TRUE.}

\item{k}{Neighborhood size for k-NN. Will only be used when useKNN is TRUE.}

\item{useKNN}{Should k-NN be used to determine the mean distance h?}

\item{kernelType}{Type of the kernel to be used, either "gaussian" or "flat"}

\item{eps}{Stopping criterion for convergence.}

\item{maxIter}{Maximum number of iterations}

\item{lambda}{Lambda, only used when KernelType is flat}
}
\value{
List of selected features.
}
\description{
Rdrags implements the high-dimensional feature selection algorithm called
Dense Relevant Attribute Group Selector
by Yu, Ding and Loscalzo, KDD 2008

Dense Relevant Attribute Group Selector
}
\section{Rdrags functions}{

Rdrags
}

\examples{

}
